unlpimage/requirements.txt
matplotlib==3.7.1
notebook==6.5.4
numpy==1.24.3
pandas==2.0.1
Pillow==9.5.0
PySimpleGUI==4.60.4